Move getNow.js to luxon dir
