+++ app/scripts/actions/就業年報画面に今年ボタンを追加.js
@@ -1,5 +1,5 @@
 import getStartOfYear from '../modules/getStartOfYear'
-import getNow from '../modules/getNow'
+import getNow from '../modules/luxon/getNow'
 
 export default function () {
   const searchButton = getSearchButtonElement()
