+++ app/scripts/actions/就業週報月報画面に今月ボタンを追加.js
@@ -1,4 +1,4 @@
-import getNow from '../modules/getNow'
+import getNow from '../modules/luxon/getNow'
 
 export default function () {
   const searchButton = getSearchButtonElement()
